src/gemini-claude-integration/requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
